Batch_6_January2022/Week6/kodlar/deploymentWflask-nginx-uwsgi/flask/requirements.txt
Flask==2.0.1
numpy==1.19.0
opencv_python==4.5.1.48
Pillow==9.0.0
uwsgi==2.0.20
Werkzeug==2.0.1
